refactor(desktop): extract editor actions to separate module
- Move registerCopyPathLineAction and registerSaveCommand to editor-actions.ts
- Simplify editor ref type annotation
- Clean up DiffViewer component
